Expand URIs in reader
This expands relative and prefixed URIs in the reader on the stack, rather than
passing them to the caller to be dealt with.  This pushes these context-full
forms to the edge of the system as much as possible to minimise the headaches
they can cause.

Towards having stricter guarantees about nodes and eliminating the CURIE node
type altogether.

+++ src/reader.h
@@ -166,10 +166,7 @@ push_byte(SerdReader* reader, SerdNode* node, const int c)
 }
 
 static inline SerdStatus
-push_bytes(SerdReader*    reader,
-           SerdNode*      ref,
-           const uint8_t* bytes,
-           unsigned       len)
+push_bytes(SerdReader* reader, SerdNode* ref, const uint8_t* bytes, size_t len)
 {
   const bool has_space = reader->stack.buf_size >= reader->stack.size + len;
   if (has_space) {
